+++ fixed_point/lib/test/fixed_point.test.cpp
@@ -19,7 +19,7 @@ TEST(FixedPointTest, SignIsConserved)
 
 TEST(FixedPointTest, FractionalIsConserved)
 {
-    float input = -167.123456789123456;
+    float input = -static_cast<float>(167.123456789123456);
     float epsilon = 1.f / static_cast<float>(1 << 16);
 
     FixedPoint<16, 16> fixed(input);
@@ -29,8 +29,8 @@ TEST(FixedPointTest, FractionalIsConserved)
 
 TEST(FixedPointTest, AdditionWorks)
 {
-    float a = 167.123456789123456;
-    float b = 183.678912389502341;
+    float a = static_cast<float>(167.123456789123456);
+    float b = static_cast<float>(183.678912389502341);
     float epsilon = 1.f / static_cast<float>(1 << 16);
 
     FixedPoint<16, 16> aFixed(a);
@@ -59,8 +59,8 @@ TEST(FixedPointTest, AdditionOverflowWorks)
 
 TEST(FixedPointTest, SubstractionWorks)
 {
-    float a = 167.123456789123456;
-    float b = 183.678912389502341;
+    float a = static_cast<float>(167.123456789123456);
+    float b = static_cast<float>(183.678912389502341);
     float epsilon = 1.f / static_cast<float>(1 << 16);
 
     FixedPoint<16, 16> aFixed(a);
@@ -89,7 +89,7 @@ TEST(FixedPointTest, SubstractionOverflowWorks)
 
 TEST(FixedPointTest, LeftShiftWorks)
 {
-    float a = -167.123456789123456;
+    float a = -static_cast<float>(167.123456789123456);
     unsigned shiftValue = 2;
     float epsilon = 1.f / static_cast<float>(1 << 16);
 
@@ -97,7 +97,7 @@ TEST(FixedPointTest, LeftShiftWorks)
     FixedPoint<13, 16> bFixed = aFixed << shiftValue;
 
     float result = bFixed.toFloat();
-    float expected = a * (1 << shiftValue);
+    float expected = a * static_cast<float>(1 << shiftValue);
     EXPECT_NEAR(result, expected, epsilon);
 }
 
@@ -117,7 +117,7 @@ TEST(FixedPointTest, LeftShiftOverflowWorks)
 
 TEST(FixedPointTest, RightShiftWorks)
 {
-    float a = -167.123456789123456;
+    float a = -static_cast<float>(167.123456789123456);
     unsigned shiftValue = 2;
     float epsilon = 1.f / static_cast<float>(1 << 16);
 
@@ -131,7 +131,7 @@ TEST(FixedPointTest, RightShiftWorks)
 
 TEST(FixedPointTest, FixedPointNegationWorks)
 {
-    float a = 167.123456789123456;
+    float a = static_cast<float>(167.123456789123456);
     float epsilon = 1.f / static_cast<float>(1 << 16);
 
     FixedPoint<16, 16> aFixed(a);
